// src/api/generated/models/createSubjectDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Class Compass API
 * An API for the Class Compass application
 * OpenAPI spec version: 1.0
 */

/**
 * The data required to create a new subject
 */
export interface CreateSubjectDto {
	/** The subject's name */
	name: string;
	/** The subject's associated school */
	schoolId: string;
	/** The subject's teacher identifiers */
	teacherIds?: string[];
}